Added Pairing Support
Added support for pairing Lightwave RF remotes and switches with the
Arduino. Up to 6 switches can be stored in the EEPROM, there is also
functions for removing the pairing.

Added a new example to demonstrate the pairing functionality. Connect a
button to pin 4 to pair, pin 5 to clear. pin 9 is PWM'd